% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{generateClearDataSet}
\alias{generateClearDataSet}
\title{Generate Clear dataset}
\usage{
generateClearDataSet(connectionDetails, drug_list = NA, labtest_list = NA,
  date_from = "2001-01-01", date_to = "2010-03-31")
}
\value{
nothing
}
\description{
Generate Clear dataset
}
\details{
initial testing
}

