% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{getClearResultDataSet}
\alias{getClearResultDataSet}
\title{Get Clear result dataset}
\usage{
getClearResultDataSet(connectionDetails)
}
\value{
a data frame from summary table
}
\description{
Get Clear result dataset
}
\details{
initial testing
}

