% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{getDataForMatching}
\alias{getDataForMatching}
\title{Get result data set of Clear for matching}
\usage{
getDataForMatching(connectionDetails)
}
\arguments{
\item{connectionDetails}{connectionDetails information
generated by DatabaseConnector::createConnectionDetails function}
}
\value{
A data frame
generated by summary data set for matching
}
\description{
Get result data set of Clear for matching
}

