% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{runCLogit}
\alias{runCLogit}
\title{Run conditional logistic regression}
\usage{
runCLogit(matched.data)
}
\arguments{
\item{matched.data}{A data frame
generated by runMatching/runMatchingWithData function}
}
\value{
A data frame
outcome of conditional logistic regression
}
\description{
Run conditional logistic regression
}

