% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{runMatching}
\alias{runMatching}
\title{Run exact 1:4 matching}
\usage{
runMatching(connectionDetails)
}
\arguments{
\item{connectionDetails}{connectionDetails information
generated by DatabaseConnector::createConnectionDetails function}
}
\value{
A data frame
outcome of 1:4 matching adjusted for age and sex
}
\description{
Run exact 1:4 matching
}

