% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{runMatchingWithData}
\alias{runMatchingWithData}
\title{Run exact 1:4 matching with data}
\usage{
runMatchingWithData(match)
}
\value{
a matched data frame
}
\description{
Run exact 1:4 matching with data
}
\details{
initial testing
}

