% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{getClearResultDataSet}
\alias{getClearResultDataSet}
\title{Get result data set of Clear}
\usage{
getClearResultDataSet(connectionDetails)
}
\arguments{
\item{connectionDetails}{connectionDetails information
generated by DatabaseConnector::createConnectionDetails function}
}
\value{
a data frame from summary table
generated by Clear algorithm
}
\description{
Summary data set of Clear algorithm
}

