% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{getDataForMatching}
\alias{getDataForMatching}
\title{Get Clear result dataset for matching}
\usage{
getDataForMatching(connectionDetails)
}
\value{
a data frame from summary table for matching
}
\description{
Get Clear result dataset for matching
}
\details{
initial testing
}

