% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{runCLogit}
\alias{runCLogit}
\title{Run Conditional Logistic regression}
\usage{
runCLogit(matched.data)
}
\value{
a outcome of regression
}
\description{
Run Conditional Logistic regression
}
\details{
initial testing
}

