% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{runMatchingWithData}
\alias{runMatchingWithData}
\title{Run exact 1:4 matching with data}
\usage{
runMatchingWithData(match)
}
\arguments{
\item{match}{A data frame
generated by getDataForMatching function}
}
\value{
A data frame
outcome of 1:4 matching adjusted for age and sex
}
\description{
Run exact 1:4 matching with data
}

